moved fan information to FanComponent class
